Separate the Migrations and VM Migrations charts into 2 different cards
Separate the Migrations and VM Migrations charts into 2 different cards
within the Metrics tab.
