+++ packages/forklift-console-plugin/src/modules/Overview/views/overview/tabs/Metrics/cards/MigrationsChartCard.tsx
@@ -1,0 +1,180 @@
+import React, { useState } from 'react';
+import { useForkliftTranslation } from 'src/utils/i18n';
+
+import { MigrationModelGroupVersionKind, V1beta1Migration } from '@kubev2v/types';
+import { useK8sWatchResource } from '@openshift-console/dynamic-plugin-sdk';
+import { Chart, ChartAxis, ChartBar, ChartGroup, ChartTooltip } from '@patternfly/react-charts';
+import {
+  Card,
+  CardActions,
+  CardBody,
+  CardHeader,
+  CardTitle,
+  Dropdown,
+  DropdownItem,
+  Flex,
+  KebabToggle,
+} from '@patternfly/react-core';
+import chart_color_blue_200 from '@patternfly/react-tokens/dist/esm/chart_color_blue_200';
+import chart_color_green_400 from '@patternfly/react-tokens/dist/esm/chart_color_green_400';
+import chart_color_red_100 from '@patternfly/react-tokens/dist/esm/chart_color_red_100';
+
+import { TimeRangeOptions, TimeRangeOptionsDictionary } from '../utils/timeRangeOptions';
+import { MigrationDataPoint, toDataPoints } from '../utils/toDataPointsHelper';
+
+import { MigrationsCardProps } from './MigrationsCard';
+
+const toStartedMigration = (m: V1beta1Migration): string => m.status.started;
+const toFinishedMigration = (m: V1beta1Migration): string => m.status.completed;
+const toDataPointsForMigrations = (
+  allMigrations: V1beta1Migration[],
+  toTimestamp: (m: V1beta1Migration) => string,
+  selectedTimeRange: TimeRangeOptions,
+): MigrationDataPoint[] => toDataPoints(allMigrations.map(toTimestamp), selectedTimeRange);
+
+export const MigrationsChartCard: React.FC<MigrationsCardProps> = () => {
+  const { t } = useForkliftTranslation();
+  const [isDropdownOpened, setIsDropdownOpened] = useState(false);
+  const onToggle = () => setIsDropdownOpened(!isDropdownOpened);
+  const [selectedTimeRange, setSelectedTimeRange] = useState<TimeRangeOptions>(
+    TimeRangeOptions.Last7Days,
+  );
+  const [migrations] = useK8sWatchResource<V1beta1Migration[]>({
+    groupVersionKind: MigrationModelGroupVersionKind,
+    namespaced: true,
+    isList: true,
+  });
+  const migrationsDataPoints: {
+    running: MigrationDataPoint[];
+    failed: MigrationDataPoint[];
+    succeeded: MigrationDataPoint[];
+  } = {
+    running: toDataPointsForMigrations(
+      migrations.filter((m) => m?.status?.conditions?.find((it) => it?.type === 'Running')),
+      toStartedMigration,
+      selectedTimeRange,
+    ),
+    failed: toDataPointsForMigrations(
+      migrations.filter((m) => m?.status?.conditions?.find((it) => it?.type === 'Failed')),
+      toFinishedMigration,
+      selectedTimeRange,
+    ),
+    succeeded: toDataPointsForMigrations(
+      migrations.filter((m) => m?.status?.conditions?.find((it) => it?.type === 'Succeeded')),
+      toFinishedMigration,
+      selectedTimeRange,
+    ),
+  };
+
+  const maxMigrationValue = Math.max(
+    ...migrationsDataPoints.running.map((m) => m.value),
+    ...migrationsDataPoints.failed.map((m) => m.value),
+    ...migrationsDataPoints.succeeded.map((m) => m.value),
+  );
+
+  return (
+    <Card>
+      <CardHeader>
+        <CardTitle className="forklift-title">
+          {t(TimeRangeOptionsDictionary[selectedTimeRange].migrationsLabelKey)}
+        </CardTitle>
+        <CardActions>
+          <Dropdown
+            toggle={<KebabToggle onToggle={onToggle} />}
+            isOpen={isDropdownOpened}
+            isPlain
+            dropdownItems={[
+              <DropdownItem
+                onClick={() => {
+                  onToggle();
+                  setSelectedTimeRange(TimeRangeOptions.Last7Days);
+                }}
+                key="7days"
+              >
+                {t('7 days')}
+              </DropdownItem>,
+              <DropdownItem
+                onClick={() => {
+                  onToggle();
+                  setSelectedTimeRange(TimeRangeOptions.Last31Days);
+                }}
+                key="31days"
+              >
+                {t('31 days')}
+              </DropdownItem>,
+              <DropdownItem
+                onClick={() => {
+                  onToggle();
+                  setSelectedTimeRange(TimeRangeOptions.Last24H);
+                }}
+                key="24hours"
+              >
+                {t('24 hours')}
+              </DropdownItem>,
+            ]}
+          />
+        </CardActions>
+      </CardHeader>
+      <CardBody className="forklift-status-migration-chart">
+        <Flex justifyContent={{ default: 'justifyContentSpaceBetween' }}>
+          <Chart
+            ariaDesc="Bar chart with migration statistics"
+            colorScale={[
+              chart_color_blue_200.var,
+              chart_color_red_100.var,
+              chart_color_green_400.var,
+            ]}
+            domainPadding={{ x: [30, 25] }}
+            maxDomain={{ y: maxMigrationValue ? undefined : 5 }}
+            legendData={[
+              { name: t('Running'), symbol: { fill: chart_color_blue_200.var } },
+              { name: t('Failed'), symbol: { fill: chart_color_red_100.var } },
+              { name: t('Succeeded'), symbol: { fill: chart_color_green_400.var } },
+            ]}
+            legendPosition="bottom"
+            height={400}
+            width={1100}
+            padding={{
+              bottom: 75,
+              left: 50,
+              right: 30,
+              top: 50,
+            }}
+          >
+            <ChartAxis />
+            <ChartAxis dependentAxis showGrid />
+            <ChartGroup offset={11} horizontal={false}>
+              <ChartBar
+                data={migrationsDataPoints.running.map(({ dateLabel, value }) => ({
+                  x: dateLabel,
+                  y: value,
+                  name: t('Running'),
+                  label: t('{{dateLabel}} Running: {{value}}', { dateLabel, value }),
+                }))}
+                labelComponent={<ChartTooltip constrainToVisibleArea />}
+              />
+              <ChartBar
+                data={migrationsDataPoints.failed.map(({ dateLabel, value }) => ({
+                  x: dateLabel,
+                  y: value,
+                  name: t('Failed'),
+                  label: t('{{dateLabel}} Failed: {{value}}', { dateLabel, value }),
+                }))}
+                labelComponent={<ChartTooltip constrainToVisibleArea />}
+              />
+              <ChartBar
+                data={migrationsDataPoints.succeeded.map(({ dateLabel, value }) => ({
+                  x: dateLabel,
+                  y: value,
+                  name: 'Succeeded',
+                  label: t('{{dateLabel}} Succeeded: {{value}}', { dateLabel, value }),
+                }))}
+                labelComponent={<ChartTooltip constrainToVisibleArea />}
+              />
+            </ChartGroup>
+          </Chart>
+        </Flex>
+      </CardBody>
+    </Card>
+  );
+};
