+++ packages/forklift-console-plugin/src/modules/Overview/views/overview/tabs/Metrics/cards/index.ts
@@ -1,4 +1,5 @@
 // @index(['./*', /style/g], f => `export * from '${f.path}';`)
-export * from './ChartsCard';
 export * from './MigrationsCard';
+export * from './MigrationsChartCard';
+export * from './VmMigrationsChartCard';
 // @endindex
